move mock numpy functions to its own file
