+++ numba_cuda/numba/cuda/tests/nrt/test_nrt.py
@@ -3,44 +3,15 @@ import numpy as np
 import unittest
 from unittest.mock import patch
 from numba.core.runtime import rtsys
-from numba.tests.support import TestCase, EnableNRTStatsMixin
+from numba.tests.support import EnableNRTStatsMixin
 from numba.cuda.testing import CUDATestCase
+
+from .mock_numpy import cuda_empty
 
 from numba import cuda
 
-from numba.core import errors, types
-from numba.core.extending import overload
-from numba.np.arrayobj import (_check_const_str_dtype, is_nonelike,
-                               ty_parse_dtype, ty_parse_shape, numpy_empty_nd)
 
-
-def cuda_empty(shape, dtype):
-    pass
-
-
-@overload(cuda_empty)
-def ol_cuda_empty(shape, dtype):
-    _check_const_str_dtype("empty", dtype)
-    if (dtype is float or
-        (isinstance(dtype, types.Function) and dtype.typing_key is float) or
-            is_nonelike(dtype)): #default
-        nb_dtype = types.double
-    else:
-        nb_dtype = ty_parse_dtype(dtype)
-
-    ndim = ty_parse_shape(shape)
-    if nb_dtype is not None and ndim is not None:
-        retty = types.Array(dtype=nb_dtype, ndim=ndim, layout='C')
-
-        def impl(shape, dtype):
-            return numpy_empty_nd(shape, dtype, retty)
-        return impl
-    else:
-        msg = f"Cannot parse input types to function np.empty({shape}, {dtype})"
-        raise errors.TypingError(msg)
-
-
-class TestNrtRefCt(EnableNRTStatsMixin, TestCase):
+class TestNrtRefCt(EnableNRTStatsMixin, CUDATestCase):
 
     def setUp(self):
         # Clean up any NRT-backed objects hanging in a dead reference cycle
